fix(desktop): stop spurious folder picker on settings → dashboard nav
The folder-first import flow was plumbed through a zustand counter plus a
useEffect in AddRepositoryModals. The counter outlives component mounts,
but the effect runs on every mount — so once a user ever clicked "Import
existing folder," every subsequent remount of the dashboard layout
(including the one that happens when navigating back from Settings)
re-ran the effect and re-opened the native directory picker.

Remove the indirection. DashboardSidebarHeader now owns useFolderFirstImport
and renders FolderFirstImportModal itself; the dropdown item calls
folderImport.start() directly on click. No store pulse, no effect.
