+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/components/DashboardSidebar/components/DashboardSidebarHeader/DashboardSidebarHeader.tsx
@@ -6,6 +6,7 @@ import {
 	DropdownMenuItem,
 	DropdownMenuTrigger,
 } from "@superset/ui/dropdown-menu";
+import { toast } from "@superset/ui/sonner";
 import { Tooltip, TooltipContent, TooltipTrigger } from "@superset/ui/tooltip";
 import { cn } from "@superset/ui/utils";
 import { useMatchRoute, useNavigate } from "@tanstack/react-router";
@@ -21,13 +22,12 @@ import {
 import { GATED_FEATURES, usePaywall } from "renderer/components/Paywall";
 import { useCurrentPlan } from "renderer/hooks/useCurrentPlan";
 import { useHotkeyDisplay } from "renderer/hotkeys";
+import { FolderFirstImportModal } from "renderer/routes/_authenticated/_dashboard/components/AddRepositoryModals/components/FolderFirstImportModal";
+import { useFolderFirstImport } from "renderer/routes/_authenticated/_dashboard/components/AddRepositoryModals/hooks/useFolderFirstImport";
 import { OrganizationDropdown } from "renderer/routes/_authenticated/_dashboard/components/TopBar/components/OrganizationDropdown";
 import { useTasksFilterStore } from "renderer/routes/_authenticated/_dashboard/tasks/stores/tasks-filter-state";
 import { STROKE_WIDTH_THICK } from "renderer/screens/main/components/WorkspaceSidebar/constants";
-import {
-	useOpenNewProjectModal,
-	useTriggerFolderImport,
-} from "renderer/stores/add-repository-modal";
+import { useOpenNewProjectModal } from "renderer/stores/add-repository-modal";
 import { useOpenNewWorkspaceModal } from "renderer/stores/new-workspace-modal";
 
 interface DashboardSidebarHeaderProps {
@@ -39,7 +39,21 @@ export function DashboardSidebarHeader({
 }: DashboardSidebarHeaderProps) {
 	const openModal = useOpenNewWorkspaceModal();
 	const openNewProject = useOpenNewProjectModal();
-	const triggerFolderImport = useTriggerFolderImport();
+	const folderImport = useFolderFirstImport({
+		onSuccess: () => {
+			toast.success("Project ready — open it from the sidebar.");
+		},
+		onError: (message) => {
+			toast.error(`Import failed: ${message}`);
+		},
+	});
+	const handleImportFolder = () => {
+		folderImport.start().catch((err) => {
+			toast.error(
+				`Import failed: ${err instanceof Error ? err.message : String(err)}`,
+			);
+		});
+	};
 	const shortcutText = useHotkeyDisplay("NEW_WORKSPACE").text;
 	const navigate = useNavigate();
 	const matchRoute = useMatchRoute();
@@ -139,7 +153,7 @@ export function DashboardSidebarHeader({
 							<HiMiniPlus className="size-4" />
 							New project
 						</DropdownMenuItem>
-						<DropdownMenuItem onSelect={triggerFolderImport}>
+						<DropdownMenuItem onSelect={handleImportFolder}>
 							<LuFolderInput className="size-4" />
 							Import existing folder
 						</DropdownMenuItem>
@@ -180,6 +194,12 @@ export function DashboardSidebarHeader({
 						New Workspace ({shortcutText})
 					</TooltipContent>
 				</Tooltip>
+
+				<FolderFirstImportModal
+					state={folderImport.state}
+					onCancel={folderImport.cancel}
+					onConfirmCreateAsNew={folderImport.confirmCreateAsNew}
+				/>
 			</div>
 		);
 	}
@@ -210,7 +230,7 @@ export function DashboardSidebarHeader({
 							<HiMiniPlus className="size-4" />
 							New project
 						</DropdownMenuItem>
-						<DropdownMenuItem onSelect={triggerFolderImport}>
+						<DropdownMenuItem onSelect={handleImportFolder}>
 							<LuFolderInput className="size-4" />
 							Import existing folder
 						</DropdownMenuItem>
@@ -278,6 +298,12 @@ export function DashboardSidebarHeader({
 					{shortcutText}
 				</span>
 			</button>
+
+			<FolderFirstImportModal
+				state={folderImport.state}
+				onCancel={folderImport.cancel}
+				onConfirmCreateAsNew={folderImport.confirmCreateAsNew}
+			/>
 		</div>
 	);
 }
